bugfix: updated response status code
