+++ api/utils/success_response.py
@@ -31,3 +31,18 @@ def auth_response(status_code: int, message: str, access_token: str, data: Optio
         response_data["data"] = data
 
     return JSONResponse(status_code=status_code, content=jsonable_encoder(response_data))
+
+
+def fail_response(status_code: int, message: str, data: Optional[dict] = None):
+    '''Returns a JSON response for success responses'''
+
+    response_data = {
+        "status_code": status_code,
+        "success": False,
+        "message": message
+    }
+    
+    if data is not None:
+        response_data["data"] = data
+
+    return JSONResponse(status_code=status_code, content=jsonable_encoder(response_data))
